+++ R/rsp.info.R
@@ -1,15 +1,13 @@
-#' @name sp.info
+#' @name rsp.info
 #' @title re(SPECIATE) information
-#' @aliases sp_info sp_profile_info sp_species_info sp_find_profile
-#' sp_find_species
+#' @aliases rsp_info rsp_profile_info rsp_species_info rsp_find_profile
+#' rsp_find_species
 
 ###########################
 #keep think about the names
 ###########################
-#  sp_profile_info used to be sp_find_profile
-#  sp_species_info used to be sp_find_species
-#  both sp_find_ functions are currently sp_info_ wrappers
-#  should remove at some point???
+#  rsp_profile_info/ rsp_find_profile
+#  rsp_species_info/ rsp_find_species
 
 #########################
 #to think about
@@ -20,36 +18,36 @@
 
 #' @description Functions that provide (re)SPECIATE
 #' source information.
-#' \code{sp_info} generates a brief version report for the currently installed
+#' \code{rsp_info} generates a brief version report for the currently installed
 #' (re)SPECIATE data sets.
-#' \code{sp_profile_info} searches the currently installed (re)SPECIATE
+#' \code{rsp_profile_info} searches the currently installed (re)SPECIATE
 #' data sets for profile records.
-#' \code{sp_species_info} searches the currently installed (re)SPECIATE
+#' \code{rsp_species_info} searches the currently installed (re)SPECIATE
 #' data sets for species records.
 
 #' @param ... character(s), any search term(s) to use when searching
 #' the local (re)SPECIATE archive for relevant records using
-#' \code{sp_profile_info} or \code{sp_species_info}.
+#' \code{rsp_profile_info} or \code{rsp_species_info}.
 #' @param by character, the section of the archive to
-#' search, by default \code{'keywords'} for \code{sp_profile_info} and
+#' search, by default \code{'keywords'} for \code{rsp_profile_info} and
 #' \code{'species_names'} for \code{sp_species_info}.
 #' @param partial logical, if \code{TRUE} (default)
-#' \code{sp_profile_info} or \code{sp_profile_info} use partial matching.
+#' \code{rsp_profile_info} or \code{rsp_profile_info} use partial matching.
 
-#' @return \code{sp_info} provides a brief version information report on the
+#' @return \code{rsp_info} provides a brief version information report on the
 #' currently installed (re)SPECIATE archive.
-#' @return \code{sp_profile_info} returns a \code{data.frame} of
+#' @return \code{rsp_profile_info} returns a \code{data.frame} of
 #' profile information, as a \code{respeciate} object.
-#' \code{sp_species_info} returns a \code{data.frame} of
+#' \code{rsp_species_info} returns a \code{data.frame} of
 #' species information as a \code{respeciate} object.
 
 #' @examples \dontrun{
 #' profile <- "Ethanol"
-#' pr <- sp_find_profile(profile)
+#' pr <- rsp_find_profile(profile)
 #' pr
 #'
 #' species <- "Ethanol"
-#' sp <- sp_find_species(species)
+#' sp <- rsp_find_species(species)
 #' sp}
 #'
 
@@ -58,7 +56,7 @@
 
 
 #######################
-#sp_info
+#rsp_info
 #######################
 
 # tidy output???
@@ -70,10 +68,10 @@
 # this is not currently catchable!!!!
 # a <- sp_info() #a = NULL
 
-#' @rdname sp.info
+#' @rdname rsp.info
 #' @export
 
-sp_info <- function() {
+rsp_info <- function() {
   #extract profile info from archive
   .ver <- "source: SPECIATE 5.2\n\t[in (re)SPECIATE since 0.2.0]"
   .ver <- paste(.ver, "\n\t[now (re)SPECIATE ", packageVersion("respeciate"), "]", sep="")
@@ -83,10 +81,10 @@ sp_info <- function() {
 }
 
 
-#' @rdname sp.info
+#' @rdname rsp.info
 #' @export
 
-sp_profile_info <- function(..., by = "keywords", partial = TRUE) {
+rsp_profile_info <- function(..., by = "keywords", partial = TRUE) {
   #extract profile info from archive
   out <- sysdata$PROFILES
   terms <- c(...)
@@ -112,7 +110,7 @@ sp_profile_info <- function(..., by = "keywords", partial = TRUE) {
     species <- sysdata$SPECIES
     ref <- out$PROFILE_CODE
     for(ti in terms){
-      ans <- sp_find_species(ti, by=by, partial=partial)
+      ans <- rsp_species_info(ti, by=by, partial=partial)
       terms <- species$PROFILE_CODE[species$SPECIES_ID %in% ans$SPECIES_ID]
       ref <- ref[ref %in% terms]
     }
@@ -129,23 +127,24 @@ sp_profile_info <- function(..., by = "keywords", partial = TRUE) {
       }
     }
   }
-  out <- rsp_build_respeciate(out)
+  out <- .rsp_build_respeciate(out)
+  class(out) <- unique(c("rsp_pi", class(out)))
   return(out)
 }
 
-#' @rdname sp.info
+#' @rdname rsp.info
 #' @export
 
 #wrapper for above
 
-sp_find_profile <- function(...){
-  sp_profile_info(...)
+rsp_find_profile <- function(...){
+  rsp_profile_info(...)
 }
 
-#' @rdname sp.info
+#' @rdname rsp.info
 #' @export
 
-sp_species_info <- function(..., by = "species_name", partial = TRUE) {
+rsp_species_info <- function(..., by = "species_name", partial = TRUE) {
   #extract species info from archive
   out <- sysdata$SPECIES_PROPERTIES
   terms <- c(...)
@@ -160,15 +159,16 @@ sp_species_info <- function(..., by = "species_name", partial = TRUE) {
     }
   }
   #out <- PROFILES[grep(term, PROFILES[[by]], ignore.case = TRUE), ]
-  out <- rsp_build_respeciate(out)
+  out <- .rsp_build_respeciate(out)
+  class(out) <- unique(c("rsp_si", class(out)))
   return(out)
 }
 
-#' @rdname sp.info
+#' @rdname rsp.info
 #' @export
 
 #wrapper for above
 
-sp_find_species <- function(...){
-  sp_species_info(...)
+rsp_find_species <- function(...){
+  rsp_species_info(...)
 }